refactor : 구매부분 로직 분리
